Go filter package refactoring
Split filter package into sub-ones where each item corresponds to a
C filter attribute structure - IP4 network, port range and so on.
A filter attribute package provides API for instantiating Go proxy
structure and routines for building C-filter items.
