+++ modules/acl/controlplane/service.go
@@ -8,7 +8,12 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	"github.com/yanet-platform/yanet2/common/go/filter"
+	"github.com/yanet-platform/yanet2/common/go/filter/device"
+	"github.com/yanet-platform/yanet2/common/go/filter/ipnet4"
+	"github.com/yanet-platform/yanet2/common/go/filter/ipnet6"
+	"github.com/yanet-platform/yanet2/common/go/filter/portrange"
+	"github.com/yanet-platform/yanet2/common/go/filter/protorange"
+	"github.com/yanet-platform/yanet2/common/go/filter/vlanrange"
 	"github.com/yanet-platform/yanet2/controlplane/ffi"
 	"github.com/yanet-platform/yanet2/modules/acl/controlplane/aclpb"
 )
@@ -44,39 +49,39 @@ func NewACLService(agent *ffi.Agent, log *zap.SugaredLogger) *ACLService {
 func convertRules(reqRules []*aclpb.Rule) ([]AclRule, error) {
 	rules := make([]AclRule, 0, len(reqRules))
 	for _, reqRule := range reqRules {
-		devices, err := filter.MakeDevices(reqRule.Devices)
+		devices, err := device.FromDevices(reqRule.Devices)
 		if err != nil {
 			return nil, err
 		}
-		vlanRanges, err := filter.MakeVlanRanges(reqRule.VlanRanges)
+		vlanRanges, err := vlanrange.FromVlanRanges(reqRule.VlanRanges)
 		if err != nil {
 			return nil, err
 		}
-		src4s, err := filter.MakeIPNet4s(reqRule.Srcs)
+		src4s, err := ipnet4.FromIPNets(reqRule.Srcs)
 		if err != nil {
 			return nil, err
 		}
-		dst4s, err := filter.MakeIPNet4s(reqRule.Dsts)
+		dst4s, err := ipnet4.FromIPNets(reqRule.Dsts)
 		if err != nil {
 			return nil, err
 		}
-		src6s, err := filter.MakeIPNet6s(reqRule.Srcs)
+		src6s, err := ipnet6.FromIPNets(reqRule.Srcs)
 		if err != nil {
 			return nil, err
 		}
-		dst6s, err := filter.MakeIPNet6s(reqRule.Dsts)
+		dst6s, err := ipnet6.FromIPNets(reqRule.Dsts)
 		if err != nil {
 			return nil, err
 		}
-		protoRanges, err := filter.MakeProtoRanges(reqRule.ProtoRanges)
+		protoRanges, err := protorange.FromProtoRanges(reqRule.ProtoRanges)
 		if err != nil {
 			return nil, err
 		}
-		srcPortRanges, err := filter.MakePortRanges(reqRule.SrcPortRanges)
+		srcPortRanges, err := portrange.FromPortRanges(reqRule.SrcPortRanges)
 		if err != nil {
 			return nil, err
 		}
-		dstPortRanges, err := filter.MakePortRanges(reqRule.DstPortRanges)
+		dstPortRanges, err := portrange.FromPortRanges(reqRule.DstPortRanges)
 		if err != nil {
 			return nil, err
 		}
