+++ modules/acl/tests/acl/acl_test.go
@@ -8,7 +8,12 @@ import (
 	"github.com/gopacket/gopacket"
 	"github.com/gopacket/gopacket/layers"
 	"github.com/stretchr/testify/require"
-	"github.com/yanet-platform/yanet2/common/go/filter"
+	"github.com/yanet-platform/yanet2/common/go/filter/device"
+	"github.com/yanet-platform/yanet2/common/go/filter/ipnet4"
+	"github.com/yanet-platform/yanet2/common/go/filter/ipnet6"
+	"github.com/yanet-platform/yanet2/common/go/filter/portrange"
+	"github.com/yanet-platform/yanet2/common/go/filter/protorange"
+	"github.com/yanet-platform/yanet2/common/go/filter/vlanrange"
 	"github.com/yanet-platform/yanet2/common/go/xpacket"
 	mock "github.com/yanet-platform/yanet2/mock/go"
 	acl "github.com/yanet-platform/yanet2/modules/acl/controlplane"
@@ -49,47 +54,47 @@ func createBasicRules() []acl.AclRule {
 		{
 			Action:     0, // PASS
 			Counter:    "",
-			Devices:    []filter.Device{{Name: defaultDeviceName}},
-			VlanRanges: []filter.VlanRange{},
-			Src4s: []filter.IPNet4{
+			Devices:    []device.Device{{Name: defaultDeviceName}},
+			VlanRanges: []vlanrange.VlanRange{},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.2.2"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.3.1"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 150, To: 450},
 				{From: 600, To: 600},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
 		{
 			Action:     1, // DENY
 			Counter:    "",
-			Devices:    []filter.Device{{Name: defaultDeviceName}},
-			VlanRanges: []filter.VlanRange{},
-			Src4s: []filter.IPNet4{
+			Devices:    []device.Device{{Name: defaultDeviceName}},
+			VlanRanges: []vlanrange.VlanRange{},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.2.99"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.3.1"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 700, To: 700},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
@@ -130,22 +135,22 @@ func createProtocolRules() []acl.AclRule {
 	return []acl.AclRule{
 		{
 			Action:  0, // PASS
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.2.2"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.3.1"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 600, To: 600},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 1538, To: 1539}, {From: 1542, To: 1543}, // TCP with SYN flag
 				{From: 1546, To: 1547}, {From: 1550, To: 1551},
 				{From: 1554, To: 1555}, {From: 1558, To: 1559},
@@ -182,22 +187,22 @@ func createProtocolRules() []acl.AclRule {
 		},
 		{
 			Action:  1, // DENY
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("0.0.0.0"), Mask: netip.MustParseAddr("0.0.0.0")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("0.0.0.0"), Mask: netip.MustParseAddr("0.0.0.0")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 1540, To: 1543}, {From: 1548, To: 1551},
 				{From: 1556, To: 1559}, {From: 1564, To: 1567},
 				{From: 1572, To: 1575}, {From: 1580, To: 1583},
@@ -218,22 +223,22 @@ func createProtocolRules() []acl.AclRule {
 		},
 		{
 			Action:  0, // PASS
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.2.2"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.3.1"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 264, To: 264}, // ICMP Echo
 			},
 		},
@@ -269,22 +274,22 @@ func createPortRangeRules() []acl.AclRule {
 	return []acl.AclRule{
 		{
 			Action:  0, // PASS
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.2.2"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.3.1"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 150, To: 450},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
@@ -310,44 +315,44 @@ func createSubnetRules() []acl.AclRule {
 	return []acl.AclRule{
 		{
 			Action:  0, // PASS
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.2.0"), Mask: netip.MustParseAddr("255.255.255.0")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.3.1"), Mask: netip.MustParseAddr("255.255.255.255")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 600, To: 600},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
 		{
 			Action:  1, // DENY
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.99.0"), Mask: netip.MustParseAddr("255.255.255.0")},
 				{Addr: netip.MustParseAddr("0.0.0.0"), Mask: netip.MustParseAddr("0.0.0.0")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("0.0.0.0"), Mask: netip.MustParseAddr("0.0.0.0")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
@@ -374,64 +379,64 @@ func createIPv6Rules() []acl.AclRule {
 	return []acl.AclRule{
 		{
 			Action:  0, // PASS
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s:   []filter.IPNet4{},
-			Dst4s:   []filter.IPNet4{},
-			Src6s: []filter.IPNet6{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s:   []ipnet4.IPNet{},
+			Dst4s:   []ipnet4.IPNet{},
+			Src6s: []ipnet6.IPNet{
 				{Addr: netip.MustParseAddr("2001:db8::1"), Mask: netip.MustParseAddr("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff")},
 			},
-			Dst6s: []filter.IPNet6{
+			Dst6s: []ipnet6.IPNet{
 				{Addr: netip.MustParseAddr("2001:db8::2"), Mask: netip.MustParseAddr("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff")},
 			},
-			SrcPortRanges: []filter.PortRange{
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 600, To: 600},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 1536, To: 1791}, // TCP
 			},
 		},
 		{
 			Action:  0, // PASS
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s:   []filter.IPNet4{},
-			Dst4s:   []filter.IPNet4{},
-			Src6s: []filter.IPNet6{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s:   []ipnet4.IPNet{},
+			Dst4s:   []ipnet4.IPNet{},
+			Src6s: []ipnet6.IPNet{
 				{Addr: netip.MustParseAddr("2001:db8::1"), Mask: netip.MustParseAddr("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff")},
 			},
-			Dst6s: []filter.IPNet6{
+			Dst6s: []ipnet6.IPNet{
 				{Addr: netip.MustParseAddr("2001:db8::2"), Mask: netip.MustParseAddr("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff")},
 			},
-			SrcPortRanges: []filter.PortRange{
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 14976, To: 14976}, // ICMPv6 Echo
 			},
 		},
 		{
 			Action:  1, // DENY
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s:   []filter.IPNet4{},
-			Dst4s:   []filter.IPNet4{},
-			Src6s: []filter.IPNet6{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s:   []ipnet4.IPNet{},
+			Dst4s:   []ipnet4.IPNet{},
+			Src6s: []ipnet6.IPNet{
 				{Addr: netip.MustParseAddr("2001:db8::99"), Mask: netip.MustParseAddr("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff")},
 			},
-			Dst6s: []filter.IPNet6{
+			Dst6s: []ipnet6.IPNet{
 				{Addr: netip.MustParseAddr("2001:db8::2"), Mask: netip.MustParseAddr("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff")},
 			},
-			SrcPortRanges: []filter.PortRange{
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 600, To: 600},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
@@ -481,85 +486,85 @@ func createOverlappingRules() []acl.AclRule {
 	return []acl.AclRule{
 		{
 			Action:  0, // PASS
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.2.0"), Mask: netip.MustParseAddr("255.255.255.254")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("0.0.0.0"), Mask: netip.MustParseAddr("0.0.0.0")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
 		{
 			Action:  1, // DENY
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.2.0"), Mask: netip.MustParseAddr("255.255.255.240")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("0.0.0.0"), Mask: netip.MustParseAddr("0.0.0.0")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
 		{
 			Action:  0, // PASS
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.2.0"), Mask: netip.MustParseAddr("255.255.255.0")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("0.0.0.0"), Mask: netip.MustParseAddr("0.0.0.0")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
 		{
 			Action:  1, // DENY
-			Devices: []filter.Device{{Name: defaultDeviceName}},
-			Src4s: []filter.IPNet4{
+			Devices: []device.Device{{Name: defaultDeviceName}},
+			Src4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("192.0.0.0"), Mask: netip.MustParseAddr("255.255.0.0")},
 			},
-			Dst4s: []filter.IPNet4{
+			Dst4s: []ipnet4.IPNet{
 				{Addr: netip.MustParseAddr("0.0.0.0"), Mask: netip.MustParseAddr("0.0.0.0")},
 			},
-			Src6s: []filter.IPNet6{},
-			Dst6s: []filter.IPNet6{},
-			SrcPortRanges: []filter.PortRange{
+			Src6s: []ipnet6.IPNet{},
+			Dst6s: []ipnet6.IPNet{},
+			SrcPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			DstPortRanges: []filter.PortRange{
+			DstPortRanges: []portrange.PortRange{
 				{From: 0, To: 65535},
 			},
-			ProtoRanges: []filter.ProtoRange{
+			ProtoRanges: []protorange.ProtoRange{
 				{From: 4352, To: 4607}, // UDP
 			},
 		},
